Demojibakefier (first public commit).
It works properly.. about ~72% of the time. And, aware of the potential
that it could be perceived as an over-engineered Pandora's box, I'll be
laughing all the way to the gates of hell with this commit (i.e., willing
to give it a try, and see where it goes from there). (＃￣ω￣)ψ__・・・