Dump PHP 5.4 from tests.
Some of the tests use ::class, which wasn't introduced until PHP 5.5 (and
will therefore always error out on PHP 5.4 and earlier). Seeing as PHP 5.4
is the oldest PHP branch supported anyhow, and seeing as we're not talking
about the latest Common Classes Package anyhow (this is v1; v2 is the
latest), and seeing as it's likely to be quite tedious to try rewriting all
the tests in such a way as to maintain the same coverage without the need
to use ::class, I'm opting instead to just dump PHP 5.4 from the tests
entirely. Much less work, our tests suddenly stop failing, and it won't
matter anymore anyway when v1 eventually goes EoL.